bot: code style
There is no 100% canonical guide to Python formatting, but there are
some suggestions, like PEP8[1] for example.

I applied some of the suggestions by hand and also used a string
formatting, that I personally find a bit more readable and which can be
spotted often in the wild as well:

Instead of:

    "Some text" + var1 + " and maybe " + var2

Use:

    "Some text %s and maybe %s" % (var1, var2)

This allows a more fine grained control over the formatting. Python has
even more ways to format strings, but the `s ... % v ...` syntax
is the most supported. More on formatting: https://pyformat.info/

[1] https://www.python.org/dev/peps/pep-0008/